topics/minimum-moves-to-spread-stones-over-grid/minimum-moves-to-spread-stones-over-grid_test.go
// Oops, forgot to write comments. Good luck, bro.
//
// @author xuejunc deerhunter0837@gmail.com
// @create 2024/7/20 下午5:17
package minimum_moves_to_spread_stones_over_grid

import "testing"

func Test_minimumMoves(t *testing.T) {
	type args struct {
		grid [][]int
	}
	tests := []struct {
		name string
		args args
		want int
	}{
		{
			"001",
			args{[][]int{
				{1, 2, 2},
				{1, 1, 0},
				{0, 1, 1}},
			},
			4,
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := minimumMoves(tt.args.grid); got != tt.want {
				t.Errorf("minimumMoves() = %v, want %v", got, tt.want)
			}
		})
	}
}
